:recycle: alert, present, presentPhotoPicker 메서드 뷰모델로 이동 (#105)
